Split UI renderer tests for the different modes in `ContactPatient`
